--- src/components/pages/Auth.tsx
@@ -1,52 +1,0 @@
-import { useState } from 'react';
-import { signInWithTwitter } from '@/packages/auth/usecase/signInWithTwitter';
-
-export default function Auth() {
-  const [loading, setLoading] = useState(false);
-  const [email, setEmail] = useState('');
-
-  const handleLogin = async () => {
-    try {
-      setLoading(true);
-      const { error, user } = await signInWithTwitter();
-      if (error) throw error;
-    } catch (error) {
-      alert(error.error_description || error.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="row flex flex-center">
-      <div className="col-6 form-widget">
-        <h1 className="header">Supabase + Next.js</h1>
-        <p className="description">
-          Sign in via magic link with your email below
-        </p>
-        <div>
-          <input
-            className="inputField"
-            type="email"
-            placeholder="Your email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-          />
-        </div>
-        <div>
-          <button
-            type="button"
-            onClick={(e) => {
-              e.preventDefault();
-              handleLogin();
-            }}
-            className="button block"
-            disabled={loading}
-          >
-            <span>{loading ? 'Loading' : 'Send magic link'}</span>
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
